requirements.txt
fastapi
pydantic
uvicorn[standard]
pandas
numpy
matplotlib
scipy
duckdb
requests
python-multipart
openai
python-dotenv
tiktoken
networkx
openpyxl
PyPDF2
Pillow
beautifulsoup4
pyarrow
jinja2
beautifulsoup4
lxml
html5lib
pytesseract
pillow
